# server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digital Courses CRUD",
    "description": "Digital Courses Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "Collection of Users"
    }
  ],
  "paths": {
    "/user": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User by username",
        "description": "Get User by username(===TG)",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "TG username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create new user in system",
        "description": "Create new user in system",
        "requestBody": {
          "description": "Remember: TG is unique!",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "New user is created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User's data",
        "description": "As of now updating the Boolean about successful courses payment",
        "requestBody": {
          "description": "Telegram Unique nickname",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User has paid for the courses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete existing collection",
        "description": "Delete existing collection by Id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of collection that we want to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection is deleted",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Users": {
        "description": "List of Users",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "User": {
        "required": [
          "firstName",
          "lastName",
          "telegram"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          }
        }
      },
      "UserEdit": {
        "required": [
          "username"
        ],
        "type": "object",
        "properties": {
          "isPaid": {
            "type": "integer",
            "example": [
              0,
              1
            ]
          },
          "username": {
            "type": "string"
          }
        }
      }
    }
  }
}